# artifacts/openapi.yaml
openapi: 3.1.0
info:
  title: КультВыезд API
  description: Сервис организации школьных культурных выездов. Данные каталога — модельные (см. поле source), внешняя интеграция
    с PRO.Культура.РФ заглушена.
  version: 1.0.0
paths:
...
  /api/v1/culture-events:
    get:
      tags:
      - catalog
      summary: List Culture Events
      operationId: list_culture_events_api_v1_culture_events_get
      parameters:
      - name: city
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: City
      - name: age
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'Возрастной ценз: 0+, 6+, 12+, 16+'
          title: Age
        description: 'Возрастной ценз: 0+, 6+, 12+, 16+'
      - name: pushkin
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Pushkin
      - name: free
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Free
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CultureEventOut'
                title: Response List Culture Events Api V1 Culture Events Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/classes/import:
    post:
      tags:
      - classes
      summary: Import Class
      operationId: import_class_api_v1_classes_import_post
      parameters:
      - name: grade
        in: query
        required: false
        schema:
          type: string
          default: '8'
          title: Grade
      - name: letter
        in: query
        required: false
        schema:
          type: string
          default: А
          title: Letter
      - name: school_number
        in: query
        required: false
        schema:
          type: string
          default: ''
          title: School Number
      - name: school_name
        in: query
        required: false
        schema:
          type: string
          default: ''
          title: School Name
      - name: teacher_name
        in: query
        required: false
        schema:
          type: string
          default: ''
          title: Teacher Name
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_import_class_api_v1_classes_import_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/classes/{class_id}/roster:
    get:
      tags:
      - classes
      summary: Class Roster
      operationId: class_roster_api_v1_classes__class_id__roster_get
      parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: integer
          title: Class Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/excursions:
    get:
      tags:
      - excursions
      summary: List Excursions
      operationId: list_excursions_api_v1_excursions_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ExcursionOut'
                type: array
                title: Response List Excursions Api V1 Excursions Get
    post:
      tags:
      - excursions
      summary: Create Excursion
      operationId: create_excursion_api_v1_excursions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExcursionCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcursionOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/excursions/{excursion_id}:
    get:
      tags:
      - excursions
      summary: Get Excursion
      operationId: get_excursion_api_v1_excursions__excursion_id__get
      parameters:
      - name: excursion_id
        in: path
        required: true
        schema:
          type: integer
          title: Excursion Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcursionOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/excursions/{excursion_id}/dashboard:
    get:
      tags:
      - excursions
      summary: Excursion Dashboard
      operationId: excursion_dashboard_api_v1_excursions__excursion_id__dashboard_get
      parameters:
      - name: excursion_id
        in: path
        required: true
        schema:
          type: integer
          title: Excursion Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/excursions/{excursion_id}/participants:
    get:
      tags:
      - excursions
      summary: Excursion Participants
      description: 'Публичный экран родителя: ученик → участие (для mini-app по контексту).'
      operationId: excursion_participants_api_v1_excursions__excursion_id__participants_get
      parameters:
      - name: excursion_id
        in: path
        required: true
        schema:
          type: integer
          title: Excursion Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Excursion Participants Api V1 Excursions  Excursion Id  Participants Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/excursions/{excursion_id}/consent:
    post:
      tags:
      - consent
      summary: Post Consent
      operationId: post_consent_api_v1_excursions__excursion_id__consent_post
      parameters:
      - name: excursion_id
        in: path
        required: true
        schema:
          type: integer
          title: Excursion Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsentRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/excursions/{excursion_id}/ticket-confirm:
    post:
      tags:
      - consent
      summary: Post Ticket Confirm
      operationId: post_ticket_confirm_api_v1_excursions__excursion_id__ticket_confirm_post
      parameters:
      - name: excursion_id
        in: path
        required: true
        schema:
          type: integer
          title: Excursion Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketConfirmRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/excursions/{excursion_id}/remind-unconfirmed:
    post:
      tags:
      - consent
      summary: Remind Unconfirmed
      operationId: remind_unconfirmed_api_v1_excursions__excursion_id__remind_unconfirmed_post
      parameters:
      - name: excursion_id
        in: path
        required: true
        schema:
          type: integer
          title: Excursion Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemindResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/excursions/{excursion_id}/publish:
    post:
      tags:
      - consent
      summary: Publish Excursion
      description: Публикация карточки выезда в чат класса MAX (инлайн-кнопка открытия приложения).
      operationId: publish_excursion_api_v1_excursions__excursion_id__publish_post
      parameters:
      - name: excursion_id
        in: path
        required: true
        schema:
          type: integer
          title: Excursion Id
      - name: chat_id
        in: query
        required: true
        schema:
          type: integer
          title: Chat Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Publish Excursion Api V1 Excursions  Excursion Id  Publish Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/students/{student_id}/link-code:
    post:
      tags:
      - consent
      summary: Create Link Code
      operationId: create_link_code_api_v1_students__student_id__link_code_post
      parameters:
      - name: student_id
        in: path
        required: true
        schema:
          type: integer
          title: Student Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkCodeOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CultureEventOut:
      properties:
        id:
          type: integer
          title: Id
        title:
          type: string
          title: Title
        venue:
          type: string
          title: Venue
        city:
          type: string
          title: City
        age_rating:
          type: string
          title: Age Rating
        event_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Event Date
        duration_min:
          type: integer
          title: Duration Min
        price:
          type: number
          title: Price
        pushkin_eligible:
          type: boolean
          title: Pushkin Eligible
        is_free:
          type: boolean
          title: Is Free
        ticket_url:
          type: string
          title: Ticket Url
        address:
          type: string
          title: Address
        description:
          type: string
          title: Description
        source:
          type: string
          title: Source
      type: object
      required:
      - id
      - title
      - venue
      - city
      - age_rating
      - event_date
      - duration_min
      - price
      - pushkin_eligible
      - is_free
      - ticket_url
      - address
      - description
      - source
      title: CultureEventOut
...